feat(ecologits): rename ProviderCarbonFootprintZone references by Hostingone - handle circular dependencies too
