+++ api/tests/integ/test_admin/test_admin_providers.py
@@ -3,7 +3,8 @@ import logging
 from fastapi.testclient import TestClient
 import pytest
 
-from api.schemas.admin.providers import CreateProvider, ProviderCarbonFootprintZone, ProviderType
+from api.domain.provider.entities import HostingZone, ProviderType
+from api.schemas.admin.providers import CreateProvider
 from api.schemas.models import ModelType
 from api.tests.integ.utils import create_router, generate_test_id, kill_openmockllm, run_openmockllm
 from api.utils.variables import EndpointRoute
@@ -52,7 +53,7 @@ class TestAdminProviders:
             key=None,  # Mock server doesn't require authentication
             timeout=10,
             model_name=model_name,
-            model_hosting_zone=ProviderCarbonFootprintZone.WOR,
+            model_hosting_zone=HostingZone.WOR,
             model_total_params=0,
             model_active_params=0,
             qos_metric=None,
@@ -72,7 +73,7 @@ class TestAdminProviders:
             key=None,  # Mock server doesn't require authentication
             timeout=10,
             model_name=model_name,
-            model_hosting_zone=ProviderCarbonFootprintZone.WOR,
+            model_hosting_zone=HostingZone.WOR,
             model_total_params=0,
             model_active_params=0,
             qos_metric=None,
