+++ api/tests/integration/endpoints/test_audiotranscriptions.py
@@ -9,12 +9,11 @@ from api.dependencies import create_audio_transcriptions_use_case_factory
 from api.domain.audio.errors import AudioFileSizeLimitExceededError
 from api.domain.model.entities import ModelType as RouterType
 from api.domain.model.errors import StatusCodeModelError, TooBusyModelError, UnknownModelError
-from api.domain.provider.entities import ProviderType
+from api.domain.provider.entities import HostingZone, ProviderType
 from api.domain.provider.errors import NoAvailableProviderError, ProviderAdapterValidationRequestError, ProviderAdapterValidationResponseError
 from api.domain.role.entities import LimitType
 from api.domain.router.errors import RouterHasNoProvidersError, RouterHasWrongTypeError, RouterNotFoundError, RouterRateLimitExceededError
 from api.domain.user.errors import UserHasInsufficientBudgetError, UserHasNoAccessToRouterError
-from api.schemas.admin.providers import ProviderCarbonFootprintZone
 from api.schemas.models import ModelType
 from api.tests.helpers import INVALID_API_KEY, create_key
 from api.tests.integration.conftest import override_global_context
@@ -64,7 +63,7 @@ class TestCreateAudioTranscriptions:
             providers=1,
             providers__type=ProviderType.VLLM,
             providers__url=DEFAULT_PROVIDER_URL,
-            providers__model_hosting_zone=ProviderCarbonFootprintZone.FRA,  # pin to an ecologits-resolvable zone (impacts now computed from transcription tokens)
+            providers__model_hosting_zone=HostingZone.FRA,  # pin to an ecologits-resolvable zone (impacts now computed from transcription tokens)
         )
         await db_session.flush()
 
@@ -100,7 +99,7 @@ class TestCreateAudioTranscriptions:
             providers=1,
             providers__type=ProviderType.VLLM,
             providers__url=DEFAULT_PROVIDER_URL,
-            providers__model_hosting_zone=ProviderCarbonFootprintZone.FRA,
+            providers__model_hosting_zone=HostingZone.FRA,
         )
         await db_session.flush()
