+++ api/tests/integration/endpoints/test_ocr.py
@@ -8,12 +8,11 @@ import respx
 from api.dependencies import create_ocr_use_case_factory
 from api.domain.model.entities import ModelType as RouterType
 from api.domain.model.errors import StatusCodeModelError, TooBusyModelError, UnknownModelError
-from api.domain.provider.entities import ProviderType
+from api.domain.provider.entities import HostingZone, ProviderType
 from api.domain.provider.errors import NoAvailableProviderError, ProviderAdapterValidationRequestError, ProviderAdapterValidationResponseError
 from api.domain.role.entities import LimitType
 from api.domain.router.errors import RouterHasNoProvidersError, RouterHasWrongTypeError, RouterNotFoundError, RouterRateLimitExceededError
 from api.domain.user.errors import UserHasInsufficientBudgetError, UserHasNoAccessToRouterError
-from api.schemas.admin.providers import ProviderCarbonFootprintZone
 from api.schemas.models import ModelType
 from api.tests.helpers import INVALID_API_KEY, create_key
 from api.tests.integration.conftest import override_global_context
@@ -61,7 +60,7 @@ class TestCreateOCR:
             providers=1,
             providers__type=ProviderType.MISTRAL,
             providers__url=DEFAULT_PROVIDER_URL,
-            providers__model_hosting_zone=ProviderCarbonFootprintZone.FRA,  # pin to an ecologits-resolvable zone (impacts now computed from OCR completion tokens)
+            providers__model_hosting_zone=HostingZone.FRA,  # pin to an ecologits-resolvable zone (impacts now computed from OCR completion tokens)
         )
         await db_session.flush()
